# openapi.yaml
openapi: 3.1.0
info:
  title: T-Shirt Store API
  version: '1.0.0'
  description: T-Shirt Store API contract. Money is always integer cents. Dates are ISO 8601 UTC.
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/ravn-112/T-shirt-Store-API/1.0.0
  - url: https://api.tshirtstore.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /products/{productId}:
    parameters:
      - { $ref: '#/components/parameters/ProductId' }
    get:
      tags: [Products]
      summary: Product detail
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Products]
      summary: Edit name/description (MANAGER)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 160 }
                description: { type: string, maxLength: 5000, nullable: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Products]
      summary: Soft delete (MANAGER)
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{productId}/active:
    parameters: [{ $ref: '#/components/parameters/ProductId' }]
    patch:
      tags: [Products]
      summary: Enable or disable the product (MANAGER)
      description: >-
        Toggles is_active, not deleted_at — the product stays visible to MANAGER either way.
        See DELETE /products/{productId} for the separate, everyone-invisible soft delete.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  isActive: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /products/{productId}/variants:
    parameters: [{ $ref: '#/components/parameters/ProductId' }]
    get:
      tags: [Variants]
      summary: Variants with real stock (MANAGER)
      parameters:
        - { name: includeInactive, in: query, schema: { type: boolean } }
        - { name: lowStock, in: query, schema: { type: boolean } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Variant' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Variants]
      summary: Create a variant (MANAGER)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVariantDto' }
      responses:
        '201':
          description: Created
          headers: { Location: { schema: { type: string } } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Variant' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /products/{productId}/variants/{variantId}:
    parameters:
      - { $ref: '#/components/parameters/ProductId' }
      - { $ref: '#/components/parameters/VariantId' }
    patch:
      tags: [Variants]
      summary: Edit price/SKU (MANAGER)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sku: { type: string, minLength: 1, maxLength: 64 }
                priceCents: { type: integer, minimum: 1 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Variant' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Variants]
      summary: Soft delete variant (MANAGER)
      description: >-
        Always succeeds, even if the variant was sold — unlike products, there is no separate
        FK-based restriction, because order_items already freezes product name, variant label
        and price independent of this row. Mirrors DELETE /products/{productId}.
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{productId}/variants/{variantId}/active:
    parameters:
      - { $ref: '#/components/parameters/ProductId' }
      - { $ref: '#/components/parameters/VariantId' }
    patch:
      tags: [Variants]
      summary: Enable or disable the variant (MANAGER)
      description: >-
        Toggles is_active, not deleted_at — same distinction as
        PATCH /products/{productId}/active.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  isActive: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /products/{productId}/variants/{variantId}/stock:
    parameters:
      - { $ref: '#/components/parameters/ProductId' }
      - { $ref: '#/components/parameters/VariantId' }
    post:
      tags: [Variants]
      summary: Manually adjust stock, by delta (MANAGER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deltaUnits]
              properties:
                deltaUnits:
                  type: integer
                  not: { const: 0 }
                  description: "Positive adds stock, negative subtracts it. Can't be 0. Use for restocking or fixing a count — a sale reduces stock on its own, see /webhooks/stripe."
                  example: -3
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  variantId: { type: string, format: uuid }
                  previousStock: { type: integer }
                  stock: { type: integer }
                  crossedLowStockThreshold: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ───────── CART / LIKES ─────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VariantId:
      name: variantId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid data
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.tshirtstore.com/errors/validation'
            title: 'Validation failed'
            status: 400
            detail: 'One or more fields are invalid'
            code: 'VALIDATION_ERROR'
            requestId: '01J9Z0000000000000000001'
            errors: [{ field: 'field', message: 'reason for the error' }]
    Unauthorized:
      description: No valid credential
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.tshirtstore.com/errors/unauthorized'
            title: 'Unauthorized'
            status: 401
            detail: 'The request could not be authenticated'
            code: 'UNAUTHORIZED'
            requestId: '01J9Z0000000000000000002'
    Forbidden:
      description: Role has no permission on this resource
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.tshirtstore.com/errors/forbidden'
            title: 'Forbidden'
            status: 403
            detail: "You don't have permission for this action"
            code: 'FORBIDDEN'
            requestId: '01J9Z0000000000000000003'
    NotFound:
      description: Doesn't exist, or isn't visible to whoever is asking
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.tshirtstore.com/errors/not-found'
            title: 'Not Found'
            status: 404
            detail: "The resource doesn't exist"
            code: 'NOT_FOUND'
            requestId: '01J9Z0000000000000000004'
    Conflict:
      description: The current state doesn't allow this operation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.tshirtstore.com/errors/conflict'
            title: 'Conflict'
            status: 409
            detail: "This operation isn't valid right now"
            code: 'CONFLICT'
            requestId: '01J9Z0000000000000000005'
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        requestId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        hasMore: { type: boolean }
...
    Color:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        hexCode: { type: string, example: '#1A1A1A' }
    Size:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: 'M' }
        position: { type: integer }
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
    ProductImage:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        altText: { type: string, nullable: true }
        position: { type: integer }
    Variant:
      type: object
      description: 'Includes real stock. Only visible to MANAGER.'
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        color: { $ref: '#/components/schemas/Color' }
        size: { $ref: '#/components/schemas/Size' }
        priceCents: { type: integer }
        stock: { type: integer }
        isActive: { type: boolean }
    PublicVariant:
      type: object
      description: 'For the public catalog and cart. Never exposes the exact stock.'
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        color: { $ref: '#/components/schemas/Color' }
        size: { $ref: '#/components/schemas/Size' }
        priceCents: { type: integer }
        isActive: { type: boolean }
        availableQuantity:
          { type: integer, example: 10, description: 'min(stock, 10)' }
        lowStock: { type: boolean, description: 'true if stock <= 3' }
    ProductCard:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        priceRange:
          type: object
          properties:
            minCents: { type: integer }
            maxCents: { type: integer }
        primaryImage: { $ref: '#/components/schemas/ProductImage' }
        categories:
          type: array
          items: { $ref: '#/components/schemas/Category' }
        inStock: { type: boolean }
        likesCount: { type: integer }
        likedByMe: { type: boolean }
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductCard'
        - type: object
          properties:
            description: { type: string, nullable: true }
            isActive: { type: boolean }
            images:
              type: array
              items: { $ref: '#/components/schemas/ProductImage' }
            variants:
              type: array
              items: { $ref: '#/components/schemas/PublicVariant' }
...
    CreateVariantDto:
      type: object
      required: [colorId, sizeId, sku, priceCents]
      properties:
        colorId: { type: string, format: uuid }
        sizeId: { type: string, format: uuid }
        sku: { type: string, minLength: 1, maxLength: 64 }
        priceCents: { type: integer, minimum: 1 }
        stock: { type: integer, minimum: 0, default: 0 }
        isActive: { type: boolean, default: true }
...